+++ src/app/connectionStrings/sagas/setConnectionStringsSaga.spec.ts
@@ -5,8 +5,25 @@
 import { call } from 'redux-saga/effects';
 import { cloneableGenerator } from 'redux-saga/utils';
 import { setConnectionStringsAction } from '../actions';
-import { setConnectionStringsSaga } from './setConnectionStringsSaga';
-import { setConnectionStrings } from './addConnectionStringSaga';
+import { getConnectionStrings, setConnectionStringsSaga, setConnectionStrings } from './setConnectionStringsSaga';
+import { CONNECTION_STRING_NAME_LIST } from '../../constants/browserStorage';
+
+describe('getConnectionString', () => {
+    it('returns expected value', () => {
+        const setValue = 'helloworld';
+        localStorage.setItem(CONNECTION_STRING_NAME_LIST, setValue);
+        expect(getConnectionStrings()).toEqual(setValue);
+    });
+});
+
+describe('setConnectionString', () => {
+    it('sets expected value', () => {
+        const setValue = 'cruelworld';
+        setConnectionStrings(setValue);
+
+        expect(localStorage.getItem(CONNECTION_STRING_NAME_LIST)).toEqual(setValue);
+    });
+});
 
 describe('setConnectionStringsSaga', () => {
     const setConnectionStringsSagaGenerator = cloneableGenerator(setConnectionStringsSaga)(setConnectionStringsAction(['connectionString2']));
